requirements.txt
streamlit
opencv-python-headless
numpy
Pillow
streamlit-drawable-canvas
pandas
openai
